Basic Notepad/Text Editor
First commit to get something going. Basic notepad/text editor setup with a couple starting pieces for a toolbar and status bar. Will add functionality to file with next commit and additionally add keyboard shortcuts such as Ctrl C/Ctrl V. 
I'm going to be working on the GUI and main elements of the app working as a text editor for a few commits. I want the app to look and run smoothly before adding any compilers/ sub processes.  My plan is to build this entire V1 in Python and PyQt with an idea of rebuilding it in C++ down the line for more speedy processing which makes a MAJOR difference in an IDE and because of my final goal with this project to launch an executable for free us which C++ makes simpler!