Compiler function!
Found a nice function on stack overflow that helped me envision how I would call compilers for other languages. I used this a blueprint and modified it to fit my use case. Currently it is untested whether it works but It will test for the OS type and check whether its Windows (.exe) or Mac/Linux (a.out) which allows for the execution of programs regardless of OS barriers. Currently just fleshing out the backend and making it a good text editor before a full IDE. 

I committed over GitHub Desktop some new updates which add functions for opening files and saving but it didn't appear to upload. I will access my other system and merge these two commits into one big one when I get to it. Currently all that needs to be worked on before the IDE aspects are QOL aspects.