Large feature overhaul
Implemented a tab system to allow multiple docs to be open at once. This required a new function to get the editor object for other events like undo, redo, cut and paste, etc. Plan to implement more key shortcuts using these.

I learned lambda functions somewhat with having to define my first one as the getEditor function and using it on the event of those previously listed. I plan to implement these more as the project goes on.

Also added a stylesheet which will allow me to further customize the layout/GUI.